pyproject.toml
[tool.poetry]
name = "telegram-notes"
version = "0.1.0"
description = ""
authors = ["reverencel <reverencel@gmail.com>"]
readme = "README.md"
packages = [{include = "telegram_notes"}]

[tool.poetry.dependencies]
python = "^3.11"
aiogram-dialog = "^2.1.0"
adaptix = "^3.0.0a8"
alembic = "^1.13.0"
aiosqlite = "^0.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
